requirements.txt
langchain
openai
python-dotenv
black[d]
isort